posix/outpost/rtos/thread.cpp: replace pthread_yield() with sched_yield()
Manual: http://man7.org/linux/man-pages/man3/pthread_yield.3.html suggests

```
This call is nonstandard, but present on several other systems.  Use
the standardized sched_yield(2) instead.
```

The pthread_yield() call is not available on macOS, but sched_yield() is
available so it seems reasonable and safe to use the sched_yield() version.

See also:

```
NOTES

On Linux, this function is implemented as a call to sched_yield(2).
```